Extracts InOutOctree BlockData classes into separate file
